pyproject.toml
[project]
name = "shade-ai"
version = "2.0.0"
description = "Fully automatic censorship removal for language models"
readme = "README.md"
license = "AGPL-3.0-or-later"
authors = [
    { name = "Assem Sabry" }
]
requires-python = ">=3.10"
keywords = ["llm", "transformer", "abliteration"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Environment :: GPU",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "accelerate~=1.10",
    "bitsandbytes~=0.45",
    "click~=8.1",
    "datasets~=4.0",
    "hf-transfer~=0.1",
    "huggingface-hub~=0.34",
    "kernels~=0.11",
    "optuna~=4.5",
    "peft~=0.14",
    "fastapi~=0.115",
    "psutil~=7.1",
    "pydantic-settings~=2.10",
    "questionary~=2.1",
    "requests~=2.32",
    "rich~=14.1",
    "transformers~=4.57",
    "uvicorn~=0.34",
]

[project.optional-dependencies]
research = [
    "geom-median~=0.1",
    "imageio~=2.37",
    "matplotlib~=3.10",
    "numpy~=2.2",
    "pacmap~=0.8",
    "scikit-learn~=1.7",
]

[dependency-groups]
dev = [
    "ruff>=0.14.5",
    "ty>=0.0.5",
]

[project.urls]
Homepage = "https://github.com/AssemSabry/Shade"
Documentation = "https://github.com/AssemSabry/Shade"
Repository = "https://github.com/AssemSabry/Shade.git"
Issues = "https://github.com/AssemSabry/Shade/issues"
Changelog = "https://github.com/AssemSabry/Shade/releases"

[project.scripts]
shade = "shade.cli:main"

[build-system]
requires = ["uv_build>=0.8.11,<0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "shade"
